% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-ru.R
\name{lang.support.ru}
\alias{lang.support.ru}
\title{Language support for Russian}
\usage{
lang.support.ru(...)
}
\arguments{
\item{...}{Optional arguments for \code{\link[koRpus:set.lang.support]{set.lang.support}}.}
}
\description{
This function adds support for Russian to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
being loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "ru" to be used with koRpus
 \item \code{treetag}: The additional preset "ru",
      implemented according to the respective TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter set[2]
}
Hyphenation patterns are provided by means of the \code{\link[sylly.ru:hyph.support.ru]{sylly.ru}} package.
}
\examples{
lang.support.ru()
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://corpus.leeds.ac.uk/mocky/ru-table.tab}
}
