% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRpus.lang.ru-package.R
\docType{package}
\name{koRpus.lang.ru-package}
\alias{koRpus.lang.ru-package}
\title{The koRpus.lang.ru Package}
\description{
Language Support for 'koRpus' Package: Russian.
}
\details{
\tabular{ll}{
Package: \tab koRpus.lang.ru\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2018-01-05\cr
Depends: \tab R (>= 3.1),koRpus (>= 0.11-2)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=koRpus\cr
}

Adds support for the Russian language to the 'koRpus' package. To ask for help, report
bugs, suggest feature improvements, or discuss the global development of the package,
      please
consider subscribing to the koRpus-dev mailing list (<http://korpusml.reaktanz.de>).
}
\author{
Meik Michalke
}
\keyword{package}
