% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-ru.R
\name{lang.support.ru}
\alias{lang.support.ru}
\title{Language support for Russian}
\usage{
lang.support.ru()
}
\description{
This function adds support for Russian to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "ru" to be used with koRpus
 \item \code{treetag}: The additional preset "ru",
      implemented according to the respective TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter set[2]
}

To also be able to use hyphenation, the package sylly.ru needs to exist and be loaded.
}
\examples{
\dontrun{
lang.support.ru()
}
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://corpus.leeds.ac.uk/mocky/ru-table.tab}
}
